Move UserDiscography to components
