+++ app/components/UserDiscography/types.ts
@@ -1,0 +1,6 @@
+import { AlbumType, TrackType } from '../Entity/types';
+
+export interface UserDiscographyProps {
+  tracks: TrackType[];
+  albums: AlbumType[];
+}
