+++ app/components/UserDiscography/userDiscography.css
@@ -1,0 +1,3 @@
+.component.userDiscography {
+  position: relative;
+}
